+++ ice-20201109/alibabacloud_ice20201109/models.py
@@ -8891,8 +8891,18 @@ class GetBatchMediaProducingJobRequest(TeaModel):
     def __init__(
         self,
         job_id: str = None,
+        signature: str = None,
+        signature_mehtod: str = None,
+        signature_nonce: str = None,
+        signature_type: str = None,
+        signature_version: str = None,
     ):
         self.job_id = job_id
+        self.signature = signature
+        self.signature_mehtod = signature_mehtod
+        self.signature_nonce = signature_nonce
+        self.signature_type = signature_type
+        self.signature_version = signature_version
 
     def validate(self):
         pass
@@ -8905,12 +8915,32 @@ class GetBatchMediaProducingJobRequest(TeaModel):
         result = dict()
         if self.job_id is not None:
             result['JobId'] = self.job_id
+        if self.signature is not None:
+            result['Signature'] = self.signature
+        if self.signature_mehtod is not None:
+            result['SignatureMehtod'] = self.signature_mehtod
+        if self.signature_nonce is not None:
+            result['SignatureNonce'] = self.signature_nonce
+        if self.signature_type is not None:
+            result['SignatureType'] = self.signature_type
+        if self.signature_version is not None:
+            result['SignatureVersion'] = self.signature_version
         return result
 
     def from_map(self, m: dict = None):
         m = m or dict()
         if m.get('JobId') is not None:
             self.job_id = m.get('JobId')
+        if m.get('Signature') is not None:
+            self.signature = m.get('Signature')
+        if m.get('SignatureMehtod') is not None:
+            self.signature_mehtod = m.get('SignatureMehtod')
+        if m.get('SignatureNonce') is not None:
+            self.signature_nonce = m.get('SignatureNonce')
+        if m.get('SignatureType') is not None:
+            self.signature_type = m.get('SignatureType')
+        if m.get('SignatureVersion') is not None:
+            self.signature_version = m.get('SignatureVersion')
         return self